tree-construction/tests2.dat

#data
<!doctype html>Test
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "Test"

#data
<textarea>test</div>test
#errors
10: missing document type declaration.
17: unescaped '</' in CDATA or RCDATA block.
25: unexpected end of file while parsing CDATA section for element textarea.
#document
| <html>
|   <head>
|   <body>
|     <textarea>
|       "test</div>test"

#data
<table><td>
#errors
7: missing document type declaration.
11: required tr element start tag implied by unexpected td element start tag.
12: unexpected end of file implied table element end tag.
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<table><td>test</tbody></table>
#errors
missing document type declarattion
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "test"

#data
<frame>test
#errors
missing document type declaration
frame element can't occur here
#document
| <html>
|   <head>
|   <body>
|     "test"

#data
<!doctype html><frameset>test
#errors
frameset can't contain text
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!doctype html><frameset><!doctype html>
#errors
document type declaration can only occur at the start of a document
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!doctype html><font><p><b>test</font>
#errors
...
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <font>
|     <p>
|       <font>
|         <b>
|           "test"

#data
<!DOCTYPE htmL><dt><div><dd>
#errors
...
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dt>
|       <div>
|     <dd>

#data
<script></x
#errors
no document type
</ in script
#document
| <html>
|   <head>
|     <script>
|       "</x"
|   <body>
